+++ src/main/java/edu/ie3/simona/api/data/primarydata/ontology/PrimaryDataMessageFromExt.java
@@ -1,0 +1,12 @@
+/*
+ * © 2024. TU Dortmund University,
+ * Institute of Energy Systems, Energy Efficiency and Energy Economics,
+ * Research group Distribution grid planning and operation
+ */
+
+package edu.ie3.simona.api.data.primarydata.ontology;
+
+import edu.ie3.simona.api.data.ontology.DataMessageFromExt;
+
+/** Messages that are sent from an external primary data simulation to SIMONA */
+public interface PrimaryDataMessageFromExt extends DataMessageFromExt {}