+++ src/main/java/edu/ie3/simona/api/exceptions/ConvertionException.java
@@ -1,0 +1,14 @@
+/*
+ * © 2024. TU Dortmund University,
+ * Institute of Energy Systems, Energy Efficiency and Energy Economics,
+ * Research group Distribution grid planning and operation
+ */
+
+package edu.ie3.simona.api.exceptions;
+
+public class ConvertionException extends Exception {
+
+  public ConvertionException(final String message) {
+    super(message);
+  }
+}